temp: update CRUD for all, auth secure, multiple changes

+++ backend/Controllers/Admin-Controllers.js
@@ -16,35 +16,29 @@ const createAdmin = async (req, res, next) => {
     });
   }
   const { firstName, lastName, email, password, mobile } = req.body;
+
   let existingAdmin;
   try {
     existingAdmin = await Admin.findOne({ email: email });
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong fetching the data, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong fetching the data, please try again", 500));
   }
   if (existingAdmin) {
-    const error = new HttpError("email already exists, please try again", 500);
-    return next(error);
+    return next(new HttpError("Email already exists, please try again", 422));
   }
+
   let hashedPassword;
   try {
     hashedPassword = await bcrypt.hash(password, 12);
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while encrypting the password, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while encrypting the password, please try again", 500));
   }
+
   if (!req.files || !req.files["image"] || req.files["image"].length === 0) {
-    const error = new HttpError("No image was uploaded.", 400);
-    return next(error);
+    return next(new HttpError("No image was uploaded.", 400));
   }
   const image = req.files["image"][0].path;
+
   const createdAdmin = new Admin({
     firstName,
     lastName,
@@ -57,30 +51,20 @@ const createAdmin = async (req, res, next) => {
   try {
     await createdAdmin.save();
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while creating admin, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while creating admin, please try again", 500));
   }
+
   let token;
   try {
     token = jwt.sign(
-      {
-        userId: createdAdmin.id,
-        email: createdAdmin.email,
-        role: createdAdmin.role,
-      },
+      { userId: createdAdmin.id, email: createdAdmin.email, role: createdAdmin.role },
       process.env.JWT_KEY,
-      { expiresIn: JWT_EXPIRY },
+      { expiresIn: JWT_EXPIRY }
     );
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while creating the JWT token, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while creating the JWT token, please try again", 500));
   }
+
   res.status(201).json({
     userId: createdAdmin.id,
     email: createdAdmin.email,
@@ -94,78 +78,59 @@ const getAllAdmins = async (req, res, next) => {
   try {
     admins = await Admin.find({}, "-password");
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while fetching the data, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the data, please try again", 500));
   }
   res.json({ admins: admins });
 };
 
 const getAdminById = async (req, res, next) => {
-  let admin;
   const id = req.params.id;
+  let admin;
   try {
     admin = await Admin.findOne({ _id: id }, "-password");
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while fetching the data, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the data, please try again", 500));
+  }
+  if (!admin) {
+    return next(new HttpError("Admin not found", 404));
   }
   res.json({ admin: admin });
 };
 
 const login = async (req, res, next) => {
   const { email, password } = req.body;
+
   let existingAdmin;
   try {
     existingAdmin = await Admin.findOne({ email: email });
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while verification of the admin, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while verification of the admin, please try again", 500));
   }
   if (!existingAdmin) {
-    const error = new HttpError("Invalid email, please try again", 401);
-    return next(error);
+    return next(new HttpError("Invalid email, please try again", 401));
   }
+
   let isValidPassword = false;
   try {
     isValidPassword = await bcrypt.compare(password, existingAdmin.password);
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while verification of the password, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while verification of the password, please try again", 500));
   }
   if (!isValidPassword) {
-    const error = new HttpError("Invalid credentials, please try again", 401);
-    return next(error);
+    return next(new HttpError("Invalid credentials, please try again", 401));
   }
+
   let token;
   try {
     token = jwt.sign(
-      {
-        userId: existingAdmin.id,
-        email: existingAdmin.email,
-        role: existingAdmin.role,
-      },
+      { userId: existingAdmin.id, email: existingAdmin.email, role: existingAdmin.role },
       process.env.JWT_KEY,
-      { expiresIn: JWT_EXPIRY },
+      { expiresIn: JWT_EXPIRY }
     );
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while fetching the JWT token, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the JWT token, please try again", 500));
   }
+
   res.status(200).json({
     userId: existingAdmin.id,
     email: existingAdmin.email,
@@ -177,75 +142,60 @@ const login = async (req, res, next) => {
 const updateAdminById = async (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new HttpError("Invalid inputs passed, please try again", 422);
-    return next(error);
+    return next(new HttpError("Invalid inputs passed, please try again", 422));
   }
+
   const id = req.params.id;
   const { firstName, lastName, password, mobile } = req.body;
+
   let admin;
   try {
     admin = await Admin.findOne({ _id: id });
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while fetching the data, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the data, please try again", 500));
   }
   if (!admin) {
-    const error = new HttpError("No admin found, please try again", 500);
-    return next(error);
+    return next(new HttpError("No admin found, please try again", 404));
   }
-  let hashedPassword;
+
   let updatedPassword;
   if (password == null) {
     updatedPassword = admin.password;
   } else {
     try {
-      hashedPassword = await bcrypt.hash(password, 12);
-      updatedPassword = hashedPassword;
+      updatedPassword = await bcrypt.hash(password, 12);
     } catch (err) {
-      const error = new HttpError(
-        "Something went wrong while encrypting the password, please try again",
-        500,
-      );
-      return next(error);
+      return next(new HttpError("Something went wrong while encrypting the password, please try again", 500));
     }
   }
-  let image = null;
-  if (req.files && req.files["image"]) {
-    image = req.files["image"][0].path;
-  }
-  admin.firstName = firstName ? firstName : admin.firstName;
-  admin.lastName = lastName ? lastName : admin.lastName;
+
+  if (firstName !== undefined) admin.firstName = firstName;
+  if (lastName !== undefined) admin.lastName = lastName;
+  if (mobile !== undefined) admin.mobile = mobile;
   admin.password = updatedPassword;
-  admin.mobile = mobile ? mobile : admin.mobile;
   admin.role = "Admin";
-  admin.image = image ? image : admin.image;
+  if (req.files && req.files["image"] && req.files["image"].length > 0) {
+    admin.image = req.files["image"][0].path;
+  }
+
   try {
     await admin.save();
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while the updation of admin, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while updating admin, please try again", 500));
   }
+
   let token;
   try {
     token = jwt.sign(
       { userId: admin.id, email: admin.email, role: admin.role },
       process.env.JWT_KEY,
-      { expiresIn: JWT_EXPIRY },
+      { expiresIn: JWT_EXPIRY }
     );
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while fetching the JWT token, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the JWT token, please try again", 500));
   }
-  res.status(201).json({
+
+  res.status(200).json({
     userId: admin.id,
     email: admin.email,
     role: admin.role,
@@ -259,21 +209,18 @@ const updateImageById = async (req, res, next) => {
   try {
     admin = await Admin.findById(id);
   } catch (err) {
-    const error = new HttpError("Something wrong while fetching the admin");
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the admin", 500));
   }
   if (!admin) {
-    const error = new HttpError("Admin not found");
-    return next(error);
+    return next(new HttpError("Admin not found", 404));
   }
   admin.image = req.file.path;
   try {
     await admin.save();
   } catch (err) {
-    const error = new HttpError("Error occured while saving the admin");
-    return next(error);
+    return next(new HttpError("Error occurred while saving the admin", 500));
   }
-  res.status(201).json({ message: "Admin updated successfully" });
+  res.status(200).json({ message: "Admin updated successfully" });
 };
 
 const deleteAdmin = async (req, res, next) => {
@@ -281,22 +228,26 @@ const deleteAdmin = async (req, res, next) => {
   let admin;
   try {
     admin = await Admin.findOne({ _id: id });
-    if (!admin) {
-      const error = new HttpError("no admin found, please try again", 500);
-      return next(error);
-    }
-    const imagePath = admin.image;
-    await admin.deleteOne();
-    res.status(200).json({ message: "Admin successfully deleted" });
-    fs.unlink(imagePath, (err) => {
-      console.log(err);
-    });
   } catch (err) {
-    const error = new HttpError(
-      "Something went wrong while deleting the admin, please try again",
-      500,
-    );
-    return next(error);
+    return next(new HttpError("Something went wrong while fetching the data, please try again", 500));
+  }
+  if (!admin) {
+    return next(new HttpError("No admin found, please try again", 404));
+  }
+
+  const imagePath = admin.image;
+  try {
+    await admin.deleteOne();
+  } catch (err) {
+    return next(new HttpError("Something went wrong while deleting the admin, please try again", 500));
+  }
+
+  res.status(200).json({ message: "Admin successfully deleted" });
+
+  if (imagePath) {
+    fs.unlink(imagePath, (err) => {
+      if (err) console.log("Image cleanup error:", err);
+    });
   }
 };
